requirements.txt
aspose-cells==23.5.0
attrs==23.1.0
colour==0.1.5
et-xmlfile==1.1.0
Jinja2==3.1.2
JPype1==1.4.1
jsonschema==4.17.3
MarkupSafe==2.1.2
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
styleframe==4.1
tqdm==4.66.1
tzdata==2023.3
XlsxWriter==3.1.2
xlwings==0.30.8
